+++ include/simuConsts.h
@@ -162,7 +162,7 @@ struct plotparameters {
 };
 
 struct numberrays {
-    int nbeams;
+    size_t nbeams;
     int number_rotati;
     int number_events;
 };
@@ -230,7 +230,7 @@ struct energy_vecs_CUDA {
     double *cum_ints;
     double *intensity_two_derivs;
     double *lamda_two_derivs;
-    int64_t size;
+    size_t size;
 };
 
 
@@ -242,10 +242,10 @@ struct plotresponc_vecs_CUDA {
     double *reflecti_two_deriv_ps;
     double *reflecti_total_ss;
     double *reflecti_two_deriv_ss;
-    int64_t size;
+    size_t size;
 
     plotresponc_vecs_CUDA (double *degrees, double *reflecti_totals, double *reflecti_two_derivs, double *reflecti_total_ps, double *reflecti_two_deriv_ps,
-                            double *reflecti_total_ss, double *reflecti_two_deriv_ss, int64_t size) :
+                            double *reflecti_total_ss, double *reflecti_two_deriv_ss, size_t size) :
                             degrees (degrees), reflecti_totals (reflecti_totals), reflecti_two_derivs (reflecti_two_derivs), reflecti_total_ps (reflecti_total_ps),
                             reflecti_two_deriv_ps (reflecti_two_deriv_ps), reflecti_total_ss (reflecti_total_ss), reflecti_two_deriv_ss (reflecti_two_deriv_ss), size (size)
     {}
